tests(Testcafe): Split list a11y checks into two fixtures
