+++ e2e/testcafe-devextreme/tests/accessibility/list/complexItems.ts
@@ -6,35 +6,6 @@ import { Options } from '../../../helpers/generateOptionMatrix';
 fixture.disablePageReloads`Accessibility`
   .page(url(__dirname, '../../container.html'));
 
-const simpleItems = ['Item_1', 'Item_2', 'Item_3'];
-
-const optionsWithSimpleItems: Options<Properties> = {
-  dataSource: [[], simpleItems],
-  height: [undefined, 400],
-  grouped: [false],
-  searchEnabled: [true, false],
-  allowItemDeleting: [true, false],
-  showSelectionControls: [true, false],
-  selectionMode: ['all', 'multiple', 'none', 'single'],
-  itemDeleteMode: ['toggle', 'context', 'slideButton', 'slideItem', 'static', 'swipe'],
-  useNativeScrolling: [true, false],
-};
-
-const a11yCheckConfig = {
-  rules: {
-    // NOTE: color-contrast issues
-    'color-contrast': { enabled: false },
-  },
-};
-
-const configurationWithSimpleItems: Configuration = {
-  component: 'dxList',
-  a11yCheckConfig,
-  options: optionsWithSimpleItems,
-};
-
-testAccessibility(configurationWithSimpleItems);
-
 const groupedItems = [
   {
     key: 'Mr. John Heart',
@@ -77,6 +48,13 @@ const options: Options<Properties> = {
   selectionMode: ['all', 'multiple', 'none', 'single'],
   itemDeleteMode: ['toggle', 'context', 'slideButton', 'slideItem', 'static', 'swipe'],
   useNativeScrolling: [true, false],
+};
+
+const a11yCheckConfig = {
+  rules: {
+    // NOTE: color-contrast issues
+    'color-contrast': { enabled: false },
+  },
 };
 
 const configuration: Configuration = {
